+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix `SkipsEmptyRows` support with the `WithColumnLimit` concern
 - formatColumn added range support
 
 ## [3.1.34] - 2021-12-2